% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sf.R
\name{theme_sf}
\alias{theme_sf}
\title{My custom theme}
\usage{
theme_sf(base_size = 12, base_family = "Arial", x_angle = 45,
  strip_bg_blank = FALSE, legend_bg_blank = TRUE)
}
\arguments{
\item{base_size}{Integer. Specify font size.}

\item{base_family}{Character. Specify font.}

\item{x_angle}{Integer. Specify angle (0-360 degree) to rotate labels of x-axis.}

\item{strip_bg_blank}{Logical. Should background of strip should be grey or blank?}

\item{legend_bg_blank}{Logical. Should background of legend should be grey or blank?}
}
\description{
My custom theme
}
